apps/py-rpa/pyproject.toml
[project]
name = "py-rpa"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "opencv-python>=4.12.0.88",
  "pillow>=11.3.0",
  "pyautogui>=0.9.54",
  "pyperclip>=1.10.0",
  "pyscreeze>=1.0.1",
  "python-dotenv>=1.1.1",
  "pywinctl>=0.4.1",
  "selenium>=4.35.0",
  "playwright>=1.49.0",
]

[dependency-groups]
dev = [
    "pyright>=1.1.405",
    "ruff>=0.13.2",
]
